refactor: moved the DoxieScan class and modernised it
